// src/lib/protocol/types.ts
export interface MissionItem {
  seq: number;
  /** MAV_FRAME enum. */
  frame: number;
  /** MAV_CMD enum. */
  command: number;
  current: number;
  autocontinue: number;
  param1: number;
  param2: number;
  param3: number;
  param4: number;
  /** Latitude * 1e7. */
  x: number;
  /** Longitude * 1e7. */
  y: number;
  /** Altitude in meters. */
  z: number;
}